+++ ChangeLog.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added `NVHPC.cmake` file for NVHPC support
+
 ## [1.1.1] 2021-11-15
 
 ### Fixed